remove networking/index
