--- template/src/app/redux/action-slice/index.ts
@@ -1,3 +1,0 @@
-export * from './app';
-
-export * from './authentication';
