+++ template/src/app/redux/store/all-reducers.ts
@@ -1,4 +1,5 @@
-import { appReducer, authenticationReducer } from '@redux-slice';
+import { appReducer } from '@redux-slice/app';
+import { authenticationReducer } from '@redux-slice/authentication';
 import { combineReducers } from '@reduxjs/toolkit';
 
 export const allReducer = combineReducers({
