+++ template/src/app/themes/index.ts
@@ -1,9 +1,4 @@
-import {
-  createStyleSheet,
-  UnistylesRegistry,
-  UnistylesThemes,
-  useStyles,
-} from 'react-native-unistyles';
+import { UnistylesRegistry, UnistylesThemes } from 'react-native-unistyles';
 
 import { darkTheme } from './dark';
 import { lightTheme } from './light';
@@ -27,5 +22,3 @@ declare module 'react-native-unistyles' {
   // eslint-disable-next-line @typescript-eslint/no-empty-interface
   export interface UnistylesThemes extends AppThemes {}
 }
-
-export { useStyles, createStyleSheet };
