+++ template/tsconfig.json
@@ -17,15 +17,13 @@
       "@components/*": ["./src/app/library/components/*"],
       "@utils/*": ["./src/app/library/utils/*"],
       "@storage": ["./src/app/library/utils/storage/index"],
-      "@networking": ["./src/app/library/networking/index"],
       "@networking/*": ["./src/app/library/networking/*"],
       "@model/*": ["./src/app/model/*"],
       "@navigation/*": ["./src/app/navigation/*"],
       "@store/*": ["./src/app/redux/store/*"],
-      "@redux-slice": ["./src/app/redux/action-slice/index"],
+      "@redux-slice/*": ["./src/app/redux/action-slice/*"],
       "@redux-selector/*": ["./src/app/redux/selector/*"],
       "@redux-action-type/*": ["./src/app/redux/action-type/*"],
-      "@theme": ["./src/app/themes/index"],
       "@theme/*": ["./src/app/themes/*"]
     }
   }
